Clean up dead code and handle IO errors
