+++ package.json
@@ -4,7 +4,6 @@
   "version": "1.0.0",
   "author": "OpenBeta <hello@openbeta.io>",
   "dependencies": {
-    "@emotion/css": "^11.1.3",
     "@mdx-js/mdx": "^1.6.22",
     "@mdx-js/react": "^1.6.22",
     "@udecode/plate": "^1.1.8",
