+++ src/components/editor/Editor.js
@@ -1,29 +1,38 @@
 import React, { useState, useEffect } from "react";
-import PlateEditor from "./PlateEditor";
 import axios from "axios";
 import queryString from "query-string";
-
 import fm from "front-matter";
 
-//TODO: make this a configurable option in gatsby-config.js
+import PlateEditor from "./PlateEditor";
+
+// TODO: make this a configurable option in gatsby-config.js
+// https://trello.com/c/K2T4B6K6
 const CONTENT_BRANCH = "develop";
 
 export const Editor = () => {
+  const [errorIO, setErrorIO] = useState(false);
   const [value, setValue] = useState(null);
   const [debug, setDebug] = useState(false);
+
   useEffect(() => {
     const get_file_from_github = async () => {
       const parsed = queryString.parse(location.search);
       if (parsed.file) {
-        const res = await client.get(
-          `${CONTENT_BRANCH}/content/${parsed.file}`
-        );
-        if (res.status === 200) {
-          const md = fm(res.data);
-          setValue(md.body);
+        try {
+          const res = await client.get(
+            `${CONTENT_BRANCH}/content/${parsed.file}`
+          );
+          if (res.status === 200) {
+            const md = fm(res.data);
+            setValue(md.body);
+          } else {
+            setErrorIO(true);
+          }
+        } catch (e) {
+          setErrorIO(true);
         }
       }
-      if (parsed.debug === 'true') {
+      if (parsed.debug === "true") {
         setDebug(true);
       }
     };
@@ -34,8 +43,22 @@ export const Editor = () => {
     console.log("## commit to github > ", markdown);
   };
 
-  return <PlateEditor markdown={value} onSubmit={onSubmit} debug={debug} />;
+  return (
+    <>
+      {errorIO && <ErrorMessage />}
+      <PlateEditor markdown={value} onSubmit={onSubmit} debug={debug} />
+    </>
+  );
 };
+
+const ErrorMessage = () => (
+  <div className="border p-4">
+    <span>Oops, something went wrong.</span>
+    <button className="btn btn-link" onClick={() => window.location.reload()}>
+      Try again
+    </button>
+  </div>
+);
 
 export const client = axios.create({
   baseURL: "https://raw.githubusercontent.com/OpenBeta/opentacos-content",
